Add SVGMapping(s) React elements
