+++ src/components/SVGMapping.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import { PanelOptionsEditorProps } from '@grafana/data';
-import { SVGIDMapping } from '../types';
 import { Button, Stack, Input, Label, Tooltip } from '@grafana/ui';
+import { SVGIDMapping } from 'types';
 
 interface SVGIDMappingProps {
   readonly value: SVGIDMapping;
@@ -11,7 +10,7 @@ interface SVGIDMappingProps {
   readonly onDelete?: (a: number) => void;
 }
 
-class SvgMapping extends React.PureComponent<SVGIDMappingProps> {
+export class SvgMapping extends React.PureComponent<SVGIDMappingProps> {
   constructor(props: SVGIDMappingProps) {
     super(props);
     this.state = { ...props.value };
@@ -73,61 +72,6 @@ class SvgMapping extends React.PureComponent<SVGIDMappingProps> {
             </Button>
           </Tooltip>
         )}
-      </Stack>
-    );
-  }
-}
-
-class SvgMappings extends React.PureComponent<PanelOptionsEditorProps<SVGIDMapping[]>> {
-  private readonly onChangeItem = (updatedMapping: SVGIDMapping, index: number): void => {
-    const newMappings = [...this.props.value];
-    newMappings[index] = updatedMapping;
-    this.props.onChange(newMappings);
-  };
-
-  private readonly onAdd = (newMapping: SVGIDMapping): void => {
-    if (newMapping.svgId !== '') {
-      const newMappings = [...this.props.value, newMapping];
-      this.props.onChange(newMappings);
-    }
-  };
-
-  private readonly onDelete = (index: number): void => {
-    const newMappings = [...this.props.value];
-    newMappings.splice(index, 1);
-    this.props.onChange(newMappings);
-  };
-
-  public render(): React.JSX.Element {
-    const svgMappings = this.props.value;
-    return (
-      <Stack direction={'column'}>
-        <Stack>
-          <Tooltip content="Clear all SVG Element ID to svgmap property mappings" theme="info">
-            <Button
-              variant="destructive"
-              icon="trash-alt"
-              size="sm"
-              onClick={() => {
-                this.props.onChange([]);
-              }}
-            >
-              Clear All
-            </Button>
-          </Tooltip>
-          <SvgMapping value={{ svgId: '', mappedName: '' }} onAdd={this.onAdd} />
-        </Stack>
-        {svgMappings.map((currentMapping: SVGIDMapping, index: number) => {
-          return (
-            <SvgMapping
-              key={currentMapping.svgId}
-              value={currentMapping}
-              index={index}
-              onChangeItem={this.onChangeItem}
-              onDelete={this.onDelete}
-            />
-          );
-        })}
       </Stack>
     );
   }
